+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1963](https://github.com/openmls/openmls/pull/1963): `MessageEncryptionError` is now part of the public API, but only surfaces if the `virtual-clients-draft` feature flag is enabled.
 - [#2043](https://github.com/openmls/openmls/pull/2043): Renamed and deprecated `MlsGroup::propose_external_psk` to `MlsGroup::propose_pre_shared_key`. Same for the `_by_value` variant.
 - [#2084](https://github.com/openmls/openmls/pull/2084): Renamed `StageCommitError::OwnCommit` to `StageCommitError::OwnCommitMismatch`, now returned only when a Commit authored by this client does not match the pending commit.
+- [#2060](https://github.com/openmls/openmls/pull/2060) Renamed `extensions-draft-08` feature flag to `extensions-draft`.
 
 ## 0.8.1 (2026-02-13)